fixresponse

+++ src/models/category.model.js
@@ -1,18 +1,22 @@
 const { sql, pool } = require("../config/db");
 
-const getAll = async (query) => {
+const getAll = async (query, userId) => {
   const { offset, limit } = query;
 
   const total = await pool
     .request()
-    .query("SELECT COUNT(id) AS total FROM Categories");
+    .input("userId", sql.Int, userId)
+    .query(
+      "SELECT COUNT(id) AS total FROM Categories Join Users  ON Categories.UserId = Users.UserId WHERE Users.UserId = @userId",
+    );
 
   const result = await pool
     .request()
+    .input("userId", sql.Int, userId)
     .input("offset", sql.Int, offset)
     .input("limit", sql.Int, limit)
     .query(
-      "SELECT id,name,description,images FROM Categories Order by id  OFFSET @offset ROWS FETCH NEXT @limit ROWS ONLY",
+      "SELECT id,name,description,images FROM Categories JOIN Users ON Categories.UserId = Users.UserId WHERE Users.UserId = @userId ORDER BY id OFFSET @offset ROWS FETCH NEXT @limit ROWS ONLY",
     );
 
   const totalpage = Math.ceil(total.recordset[0].total / limit);
@@ -35,7 +39,7 @@ const getById = async (id, userId) => {
 
   return result;
 };
-const create = async (name, description, images) => {
+const create = async (name, description, images, userId) => {
   const transaction = new sql.Transaction(pool);
   try {
     await transaction.begin();
@@ -44,8 +48,9 @@ const create = async (name, description, images) => {
       .input("name", sql.NVarChar, name)
       .input("description", sql.NVarChar, description)
       .input("images", sql.NVarChar, JSON.stringify(images))
+      .input("userId", sql.Int, userId)
       .query(
-        "INSERT INTO Categories (name,description,images) VALUES (@name, @description, @images)",
+        "INSERT INTO Categories (name,description,images,UserId) VALUES (@name, @description, @images, @userId)",
       );
     await transaction.commit();
     return [];
@@ -54,28 +59,35 @@ const create = async (name, description, images) => {
     throw error;
   }
 };
-const edit = async (id, name, State) => {
+const edit = async (id, name, State, userId) => {
   await pool
     .request()
     .input("id", sql.Int, id)
     .input("name", sql.NVarChar, name)
     .input("State", sql.VarChar, State)
-    .query("UPDATE Categories SET Name=@name, State=@State WHERE id = @id");
+    .input("userId", sql.Int, userId)
+    .query(
+      "UPDATE Categories SET Name=@name, State=@State WHERE id = @id AND UserId = @userId",
+    );
 
   return [];
 };
-const remove = async (id) => {
+const remove = async (id, userId) => {
   await pool
     .request()
     .input("id", sql.Int, id)
-    .query("DELETE FROM Categories WHERE id= @id");
+    .input("userId", sql.Int, userId)
+    .query("DELETE FROM Categories WHERE id= @id AND UserId = @userId");
   return [];
 };
-const removeimage = async (id) => {
+const removeimage = async (id, userId) => {
   await pool
     .request()
     .input("id", sql.Int, id)
-    .query("Update Categories SET images = NULL WHERE id= @id");
+    .input("userId", sql.Int, userId)
+    .query(
+      "Update Categories SET images = NULL WHERE id= @id AND UserId = @userId",
+    );
   return [];
 };
 module.exports = {
